Refactor and organize Series objs
